Added mixin, scss vars, deleted spec
