+++ src/app/app.component.scss
@@ -1,9 +1,11 @@
+@import "../styles/variables";
+
 .app-container {
   display: flex;
   flex-direction: column;
   align-items: center;
   min-height: 100vh;
-  background-color: #90ee90;
+  background-color: $color-background;
   width: 100%;
   overflow-x: hidden;
 }
