+++ src/app/pokemon-card/pokemon-card.component.scss
@@ -1,7 +1,9 @@
+@import "../../styles/variables";
+
 .pokemon-card {
   width: 200px;
   height: 200px;
-  border: 1px solid #ccc;
+  border: 1px solid $color-border;
   border-radius: 10px;
   overflow: hidden;
   box-shadow: 0 2px 5px rgba(0, 0, 0, 0.2);
@@ -11,7 +13,7 @@
   justify-content: center;
   text-align: center;
   padding: 10px;
-  background-color: #f9f9f9;
+  background-color: white;
 }
 
 .pokemon-image {
