+++ src/app/pokemon-card/pokemon-card.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input } from '@angular/core';
 
-import { Pokemon } from '../pokemon.model';
+import { Pokemon } from '../../models/pokemon.model';
 
 @Component({
   selector: 'app-pokemon-card',
