--- src/app/pokemon.model.ts
@@ -1,7 +1,0 @@
-export class Pokemon {
-  id: number;
-  name: string;
-  imageUrl: string;
-  types: string[];
-  abilities: string[];
-}
