--- src/app/pokemon.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { PokemonService } from './pokemon.service';
-
-describe('PokemonService', () => {
-  let service: PokemonService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(PokemonService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
