--- src/app/pokemons-dummy-data.ts
@@ -1,109 +1,0 @@
-import { Pokemon } from './pokemon.model';
-
-// export const DUMMY_POKEMONS: Pokemon[] = [
-//   {
-//     id: 1,
-//     name: 'Bulbasaur',
-//     imageUrl:
-//       'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/1.png',
-//     types: ['grass', 'poison'],
-//   },
-//   {
-//     id: 2,
-//     name: 'Ivysaur',
-//     imageUrl:
-//       'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/2.png',
-//     types: ['grass', 'poison'],
-//   },
-//   {
-//     id: 3,
-//     name: 'Venusaur',
-//     imageUrl:
-//       'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/3.png',
-//     types: ['grass', 'poison'],
-//   },
-//   {
-//     id: 4,
-//     name: 'Bulbasaur',
-//     imageUrl:
-//       'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/1.png',
-//     types: ['grass', 'poison'],
-//   },
-//   {
-//     id: 5,
-//     name: 'Ivysaur',
-//     imageUrl:
-//       'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/2.png',
-//     types: ['grass', 'poison'],
-//   },
-//   {
-//     id: 6,
-//     name: 'Venusaur',
-//     imageUrl:
-//       'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/3.png',
-//     types: ['grass', 'poison'],
-//   },
-//   {
-//     id: 7,
-//     name: 'Bulbasaur',
-//     imageUrl:
-//       'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/1.png',
-//     types: ['grass', 'poison'],
-//   },
-//   {
-//     id: 8,
-//     name: 'Ivysaur',
-//     imageUrl:
-//       'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/2.png',
-//     types: ['grass', 'poison'],
-//   },
-//   {
-//     id: 9,
-//     name: 'Venusaur',
-//     imageUrl:
-//       'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/3.png',
-//     types: ['grass', 'poison'],
-//   },
-//   {
-//     id: 10,
-//     name: 'Bulbasaur',
-//     imageUrl:
-//       'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/1.png',
-//     types: ['grass', 'poison'],
-//   },
-//   {
-//     id: 11,
-//     name: 'Ivysaur',
-//     imageUrl:
-//       'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/2.png',
-//     types: ['grass', 'poison'],
-//   },
-//   {
-//     id: 12,
-//     name: 'Venusaur',
-//     imageUrl:
-//       'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/3.png',
-//     types: ['grass', 'poison'],
-//   },
-//   {
-//     id: 13,
-//     name: 'Bulbasaur',
-//     imageUrl:
-//       'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/1.png',
-//     types: ['grass', 'poison'],
-//   },
-//   {
-//     id: 14,
-//     name: 'Ivysaur',
-//     imageUrl:
-//       'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/2.png',
-//     types: ['grass', 'poison'],
-//   },
-//   {
-//     id: 15,
-//     name: 'Venusaur',
-//     imageUrl:
-//       'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/3.png',
-//     types: ['grass', 'poison'],
-//   },
-// ];
