--- src/app/pokemons-list/pokemons-list.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { PokemonsListComponent } from './pokemons-list.component';
-
-describe('PokemonsListComponent', () => {
-  let component: PokemonsListComponent;
-  let fixture: ComponentFixture<PokemonsListComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [PokemonsListComponent]
-    })
-    .compileComponents();
-    
-    fixture = TestBed.createComponent(PokemonsListComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
