+++ src/models/pokemon.model.ts
@@ -1,0 +1,15 @@
+export class Pokemon {
+  id: number;
+  name: string;
+  imageUrl: string;
+  types: string[];
+  abilities: string[];
+
+  constructor() {
+    this.id = 0;
+    this.name = '';
+    this.imageUrl = '';
+    this.types = [];
+    this.abilities = [];
+  }
+}
