+++ src/styles.scss
@@ -1,4 +1,3 @@
-/* You can add global styles to this file, and also import other style files */
 html,
 body {
   font-family: "Open Sans", sans-serif;
